requirements.txt
aldryn-apphooks-config==0.6.0
aldryn-boilerplates==0.8.0
aldryn-common==1.0.5
aldryn-search==1.1.0
asgiref==3.5.2
astor==0.8.1
click==8.1.3
colorama==0.4.5
cssselect2==0.6.0
dj-database-url==1.0.0
Django==3.2.13
django-app-enabler==0.2.0
django-appconf==1.0.5
django-appdata==0.3.2
django-classy-tags==3.0.1
django-cms==3.11.0
django-filer==2.2.2
django-formtools==2.3
django-haystack==3.2.1
django-js-asset==2.0.0
django-meta==2.1.0
django-mptt==0.13.4
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.0.0
django-sortedm2m==3.1.1
django-spurl==0.6.8
django-standard-form==1.1.1
django-taggit==3.0.0
django-taggit-autosuggest==0.4.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.4
djangocms-admin-style==3.2.0
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.1.0
djangocms-blog==1.2.3
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==4.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.1
djangocms-video==3.0.0
easy-thumbnails==2.8.3
html5lib==1.1
lxml==4.9.1
Pillow==9.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tzdata==2022.1
tzlocal==4.2
Unidecode==1.1.2
URLObject==2.4.3
webencodings==0.5.1
YURL==1.0.0
